requirements.txt
streamlit
streamlit-option-menu
folium
streamlit-folium
geopandas
plotly
seaborn
matplotlib
shapely
sklearn
boruta
imblearn
xgboost
numpy
fpdf

